% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AICc}
\alias{AICc}
\alias{AICc.default}
\title{Corrected Akaike's Information Criterion}
\usage{
AICc(object, ...)
}
\arguments{
\item{object}{Time series model.}

\item{...}{Some stuff.}
}
\value{
This function returns numeric value.
}
\description{
This function extracts AICc from "smooth" objects.
}
\details{
AICc was proposed by Nariaki Sugiura in 1978 and is used on small samples.
}
\examples{

ourModel <- ces(rnorm(100,0,1),h=10)

AICc(ourModel,h=10)

}
\references{
Kenneth P. Burnham, David R. Anderson (1998). Model Selection
and Multimodel Inference. Springer Science & Business Media.
}
\seealso{
\link[stats]{AIC}, \link[stats]{BIC}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{htest}
