% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nParam}
\alias{nParam}
\alias{nParam.default}
\title{Number of parameters in the model}
\usage{
nParam(object, ...)
}
\arguments{
\item{object}{Time series model.}

\item{...}{Some other parameters passed to the method.}
}
\value{
This function returns a numeric value.
}
\description{
This function returns the number of parameters in the estimated model
}
\details{
This is a very basic and a simple function which does what it says:
extracts number of parameters in the estimated model.
}
\examples{

ourModel <- ces(rnorm(100,0,1),h=10)

nParam(ourModel)

}
\seealso{
\link[stats]{nobs}, \link[stats]{logLik}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{htest}
