% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-measures.R
\name{pinball}
\alias{pinball}
\title{Pinball function}
\usage{
pinball(holdout, forecast, level, loss = 1, digits = 5)
}
\arguments{
\item{holdout}{The vector or matrix of the holdout values.}

\item{forecast}{The forecast of prediction interval (should be the same length as the
holdout).}

\item{level}{The level of the prediction interval associated with the forecast.}

\item{loss}{The type of loss to use. The number which corresponds to L1, L2 etc.}

\item{digits}{Number of digits of the output.}
}
\value{
The function returns the scalar value.
}
\description{
The function returns the value from the pinball function for the specified level and
the type of loss
}
\examples{
y <- rnorm(100,10,2)
esmodel <- es(y,model="ANN",h=10,holdout=TRUE,intervals=TRUE,level=0.95)

# Pinball with the L1 (quantile value)
pinball(esmodel$holdout,esmodel$upper,level=0.975,loss=1,digits=5)
pinball(esmodel$holdout,esmodel$lower,level=0.025,loss=1,digits=5)

# Pinball with the L2 (expectile value)
pinball(esmodel$holdout,esmodel$upper,level=0.975,loss=2,digits=5)
pinball(esmodel$holdout,esmodel$lower,level=0.025,loss=2,digits=5)

}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{smooth}
\keyword{ts}
\keyword{univar}
