% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/msdecompose.R
\name{plot.smooth}
\alias{plot.smooth}
\alias{plot.msdecompose}
\title{Plots for the fit and states}
\usage{
\method{plot}{smooth}(x, which = c(1, 2), ask = length(which) > 1, legend = FALSE, ...)

\method{plot}{msdecompose}(x, which = c(1, 2), ask = length(which) > 1, ...)
}
\arguments{
\item{x}{Time series model for which forecasts are required.}

\item{which}{Which of the plots to produce. The possible options (see details for explanations):
\enumerate{
\item Fitted over time;
\item States of the model.
}}

\item{ask}{Logical; if \code{TRUE}, the user is asked to press Enter before each plot.}

\item{legend}{If \code{TRUE}, then the legend is produced on plots (1) and (2).}

\item{...}{The parameters passed to the plot functions. Recommended to use with separate plots.}
}
\value{
The function produces 2 plots.
}
\description{
The function produces plot actuals, fitted values and forecasts and states of the model
}
\details{
The list of produced plots includes:
\enumerate{
\item Fitted over time. Plots actuals (black line), fitted values (purple line), point forecast
and prediction interval of width \code{level}, if it was specified;
\item States of the model;
}
Which of the plots to produce, is specified via the \code{which} parameter.
}
\examples{

ourModel <- es(rnorm(100,100,10), "ANN", h=10)
plot(ourModel, c(1,2))

}
\seealso{
\link[greybox]{plot.greybox}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{ts}
\keyword{univar}
