% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pointLik}
\alias{pointLik}
\alias{pointLik.default}
\title{Point likelihood values}
\usage{
pointLik(object, ...)
}
\arguments{
\item{object}{Time series model.}

\item{...}{Some stuff.}
}
\value{
This function returns a vector.
}
\description{
This function returns a vector of logarithms of likelihoods for each observation
}
\details{
Instead of taking the expected log-likelihood for the whole series, this function calculates the individual value for each separate observation.
}
\examples{

ourModel <- ces(rnorm(100,0,1),h=10)

pointLik(ourModel,h=10)

}
\seealso{
\link[stats]{AIC}, \link[stats]{BIC}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{criteria}
\keyword{criterion}
\keyword{information}
