% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\name{refitted}
\alias{refitted}
\alias{orders}
\title{Refit the model with randomly generated initial parameters}
\usage{
refitted(object, nsim = 1000, ...)
}
\arguments{
\item{object}{Model estimated using one of the functions of smooth package.}

\item{nsim}{Number of paths to generate (number of simulations to do).}

\item{...}{Currently nothing is accepted via ellipsis}
}
\value{
The object of the class "refitted" is returns, which contains:
\itemize{
\item \code{states} - The array of states of the model;
\item \code{fitted} - The matrix with fitted values, where columns correspond
to different paths;
\item \code{transition} - The array of transition matrices;
\item \code{measurement} - The array of measurement matrices;
\item \code{persistence} - The matrix of persistence vectors (paths in columns);
\item \code{profile} - The array of profiles obtained by the end of each fit.
}
}
\description{
The functions generates the parameters based on the values in the provided object
and reapplies the same model with those parameters to the data, getting the fitted
paths.
}
\details{
The main motivation of the function is to take the randomness due to the in-sample
estimation into account when fitting the model and propagate this randomness to the
future. The method can be considered as a special case of recursive bootstrap.
}
\examples{

x <- rnorm(100,0,1)

# Just as example. orders and lags do not return anything for ces() and es(). But modelType() does.
ourModel <- adam(x, "ANN")
refittedModel <- refitted(ourModel, nsim=100)

plot(actuals(ourModel))
for(i in 1:100){lines(ourModel$fitted[,i],col="grey",lty=2)}
lines(fitted(test2),col="purple",lwd=1,lty=2)

}
\seealso{
\link[stats]{fitted}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{smooth}
\keyword{ts}
\keyword{univar}
