% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sowhat.R
\name{sowhat}
\alias{sowhat}
\title{Function returns the ultimate answer to any question}
\usage{
sowhat(...)
}
\arguments{
\item{...}{Any number of variables or string with a question.}
}
\value{
It doesn't return any value, only messages. So what?
}
\description{
You need description? So what?
}
\details{
You need details? So what?
}
\examples{

x <- rnorm(10000,0,1);
sowhat(x);

sowhat("What's the meaning of life?")

sowhat("I don't have a girlfriend.")

}
\references{
\itemize{
\item\href{https://en.wiktionary.org/wiki/so_what}{Sowhat?}
\item\href{https://www.youtube.com/watch?v=FJfFZqTlWrQ}{Sowhat?}
\item\href{https://en.wikipedia.org/wiki/Douglas_Adams}{42}
}
}
\seealso{
Nowwhat (to be implemented),
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{42}
\keyword{sowhat}
