% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-package.R
\name{ves}
\alias{ves}
\alias{viss}
\alias{sim.ves}
\title{Vector exponential smoothing}
\usage{
ves(...)

viss(...)

sim.ves(...)
}
\arguments{
\item{...}{parameters of the model}
}
\description{
VES and related functions. These have been moved to legion package starting from smooth v3.1.0.
}
