% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R, R/methods.R, R/msdecompose.R
\name{forecast.adam}
\alias{forecast.adam}
\alias{forecast.smooth}
\alias{forecast}
\alias{forecast.oes}
\alias{forecast.msdecompose}
\title{Forecasting time series using smooth functions}
\usage{
\method{forecast}{adam}(object, h = 10, newdata = NULL,
  occurrence = NULL, interval = c("none", "prediction", "confidence",
  "simulated", "approximate", "semiparametric", "nonparametric"),
  level = 0.95, side = c("both", "upper", "lower"), cumulative = FALSE,
  nsim = 10000, ...)

\method{forecast}{smooth}(object, h = 10, interval = c("parametric",
  "semiparametric", "nonparametric", "none"), level = 0.95,
  side = c("both", "upper", "lower"), ...)

\method{forecast}{oes}(object, h = 10, interval = c("parametric",
  "semiparametric", "nonparametric", "none"), level = 0.95,
  side = c("both", "upper", "lower"), ...)

\method{forecast}{msdecompose}(object, h = 10, interval = c("parametric",
  "semiparametric", "nonparametric", "none"), level = 0.95, model = NULL,
  ...)
}
\arguments{
\item{object}{Time series model for which forecasts are required.}

\item{h}{Forecast horizon.}

\item{newdata}{The new data needed in order to produce forecasts.}

\item{occurrence}{The vector containing the future occurrence variable
(values in [0,1]), if it is known.}

\item{interval}{Type of interval to construct. See \link[smooth]{es} for
details.}

\item{level}{Confidence level. Defines width of prediction interval.}

\item{side}{Defines, whether to provide \code{"both"} sides of prediction
interval or only \code{"upper"}, or \code{"lower"}.}

\item{cumulative}{If \code{TRUE}, then the cumulative forecast and prediction
interval are produced instead of the normal ones. This is useful for
inventory control systems.}

\item{nsim}{Number of iterations to do in case of \code{interval="simulated"}.}

\item{...}{Other arguments accepted by either \link[smooth]{es},
\link[smooth]{ces}, \link[smooth]{gum} or \link[smooth]{ssarima}.}

\item{model}{The type of ETS model to fit on the decomposed trend. Only applicable to
"msdecompose" class. This is then returned in parameter "esmodel". If \code{NULL}, then
it will be selected automatically based on the type of the used decomposition (either
among pure additive or among pure additive ETS models).}
}
\value{
Returns object of class "smooth.forecast", which contains:

\itemize{
\item \code{model} - the estimated model (ES / CES / GUM / SSARIMA).
\item \code{method} - the name of the estimated model (ES / CES / GUM / SSARIMA).
\item \code{forecast} aka \code{mean} - point forecasts of the model
(conditional mean).
\item \code{lower} - lower bound of prediction interval.
\item \code{upper} - upper bound of prediction interval.
\item \code{level} - confidence level.
\item \code{interval} - binary variable (whether interval were produced or not).
}
}
\description{
This function is created in order for the package to be compatible with Rob
Hyndman's "forecast" package
}
\details{
This is not a compulsory function. You can simply use \link[smooth]{es},
\link[smooth]{ces}, \link[smooth]{gum} or \link[smooth]{ssarima} without
\code{forecast.smooth}. But if you are really used to \code{forecast}
function, then go ahead!
}
\examples{

ourModel <- ces(rnorm(100,0,1),h=10)

forecast.smooth(ourModel,h=10)
forecast.smooth(ourModel,h=10,interval=TRUE)
plot(forecast.smooth(ourModel,h=10,interval=TRUE))

}
\references{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D. (2008)
Forecasting with exponential smoothing: the state space approach,
Springer-Verlag.
}
\seealso{
\code{\link[forecast]{ets}, \link[forecast]{forecast}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{ts}
\keyword{univar}
