% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-measures.R
\name{hm}
\alias{hm}
\alias{ham}
\alias{cbias}
\title{Half moment of a distribution and its derivatives.}
\usage{
hm(x, C = mean(x), digits = 5, ...)

ham(x, C = mean(x), digits = 5, ...)

cbias(x, C = mean(x), digits = 5, ...)
}
\arguments{
\item{x}{A variable based on which HM is estimated.}

\item{C}{Centering parameter.}

\item{digits}{Number of digits for rounding.}

\item{...}{Other parameters passed to mean function.}
}
\value{
A complex variable is returned for \code{hm} function and real values
are returned for \code{cbias} and \code{ham}.
}
\description{
\code{hm} function estimates half moment from some predefined constant
\code{C}. \code{ham} estimates half absolute moment. Finally, \code{cbias}
function returns bias based on \code{hm}.
}
\details{
\code{NA} values of \code{x} are excluded on the first step of calculation.
}
\examples{

x <- rnorm(100,0,1)
hm(x)
ham(x)
cbias(x)

}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{smooth}
\keyword{ts}
\keyword{univar}
