% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oes.R, R/oesg.R
\name{oes}
\alias{oes}
\title{Occurrence ETS model}
\usage{
oes(data, model = "MNN", persistence = NULL, initial = "o",
  initialSeason = NULL, occurrence = c("fixed", "general",
  "odds-ratio", "inverse-odds-ratio", "probability", "auto", "none"),
  ic = c("AICc", "AIC", "BIC", "BICc"), h = 10, holdout = FALSE,
  intervals = c("none", "parametric", "semiparametric", "nonparametric"),
  level = 0.95, bounds = c("usual", "admissible", "none"),
  silent = c("all", "graph", "legend", "output", "none"), xreg = NULL,
  xregDo = c("use", "select"), updateX = FALSE, ...)

oes(data, model = "MNN", persistence = NULL, initial = "o",
  initialSeason = NULL, occurrence = c("fixed", "general",
  "odds-ratio", "inverse-odds-ratio", "probability", "auto", "none"),
  ic = c("AICc", "AIC", "BIC", "BICc"), h = 10, holdout = FALSE,
  intervals = c("none", "parametric", "semiparametric", "nonparametric"),
  level = 0.95, bounds = c("usual", "admissible", "none"),
  silent = c("all", "graph", "legend", "output", "none"), xreg = NULL,
  xregDo = c("use", "select"), updateX = FALSE, ...)
}
\arguments{
\item{data}{Either numeric vector or time series vector.}

\item{model}{The type of ETS model used for the estimation. Normally this should
be \code{"MNN"} or any other pure multiplicative model.}

\item{persistence}{Persistence vector \eqn{g}, containing smoothing
parameters. If \code{NULL}, then estimated.}

\item{initial}{Can be either character or a vector of initial states. If it
is character, then it can be \code{"optimal"}, meaning that the initial
states are optimised, or \code{"backcasting"}, meaning that the initials are
produced using backcasting procedure.}

\item{initialSeason}{The vector of the initial seasonal components. If \code{NULL},
then it is estimated.}

\item{occurrence}{The type of model used in probability estimation. Can be
\code{"none"} - none,
\code{"fixed"} - constant probability,
\code{"odds-ratio"} - the Odds-ratio model with b=1 in Beta distribution,
\code{"inverse-odds-ratio"} - the model with a=1 in Beta distribution,
\code{"probability"} - the TSB-like (Teunter et al., 2011) probability update
mechanism a+b=1,
\code{"auto"} - the automatically selected type of occurrence model,
\code{"general"} - the general Beta model with two parameters. This will call
\code{oesg()} function with two similar ETS models and the same provided
parameters (initials and smoothing).}

\item{ic}{The information criteria to use in case of model selection.}

\item{h}{The forecast horizon.}

\item{holdout}{If \code{TRUE}, holdout sample of size \code{h} is taken from
the end of the data.}

\item{intervals}{The type of intervals to construct. This can be:

\itemize{
\item \code{none}, aka \code{n} - do not produce prediction
intervals.
\item \code{parametric}, \code{p} - use state-space structure of ETS. In
case of mixed models this is done using simulations, which may take longer
time than for the pure additive and pure multiplicative models.
\item \code{semiparametric}, \code{sp} - intervals based on covariance
matrix of 1 to h steps ahead errors and assumption of normal / log-normal
distribution (depending on error type).
\item \code{nonparametric}, \code{np} - intervals based on values from a
quantile regression on error matrix (see Taylor and Bunn, 1999). The model
used in this process is e[j] = a j^b, where j=1,..,h.
}
The parameter also accepts \code{TRUE} and \code{FALSE}. The former means that
parametric intervals are constructed, while the latter is equivalent to
\code{none}.
If the forecasts of the models were combined, then the intervals are combined
quantile-wise (Lichtendahl et al., 2013).}

\item{level}{The confidence level. Defines width of prediction interval.}

\item{bounds}{What type of bounds to use in the model estimation. The first
letter can be used instead of the whole word.}

\item{silent}{If \code{silent="none"}, then nothing is silent, everything is
printed out and drawn. \code{silent="all"} means that nothing is produced or
drawn (except for warnings). In case of \code{silent="graph"}, no graph is
produced. If \code{silent="legend"}, then legend of the graph is skipped.
And finally \code{silent="output"} means that nothing is printed out in the
console, but the graph is produced. \code{silent} also accepts \code{TRUE}
and \code{FALSE}. In this case \code{silent=TRUE} is equivalent to
\code{silent="all"}, while \code{silent=FALSE} is equivalent to
\code{silent="none"}. The parameter also accepts first letter of words ("n",
"a", "g", "l", "o").}

\item{xreg}{The vector or the matrix of exogenous variables, explaining some parts
of occurrence variable (probability).}

\item{xregDo}{Variable defines what to do with the provided xreg:
\code{"use"} means that all of the data should be used, while
\code{"select"} means that a selection using \code{ic} should be done.
\code{"combine"} will be available at some point in future...}

\item{updateX}{If \code{TRUE}, transition matrix for exogenous variables is
estimated, introducing non-linear interactions between parameters.
Prerequisite - non-NULL \code{xreg}.}

\item{...}{The parameters passed to the optimiser, such as \code{maxeval},
\code{xtol_rel}, \code{algorithm} and \code{print_level}. The description of
these is printed out by \code{nloptr.print.options()} function from the \code{nloptr}
package. The default values in the oes function are \code{maxeval=500},
\code{xtol_rel=1E-8}, \code{algorithm="NLOPT_LN_SBPLX"} and \code{print_level=0}.}

\item{phi}{The value of the damping parameter. Used only for damped-trend models.}

\item{data}{Either numeric vector or time series vector.}

\item{modelA}{The type of the ETS for the model A.}

\item{modelB}{The type of the ETS for the model B.}

\item{persistenceA}{The persistence vector \eqn{g}, containing smoothing
parameters used in the model A. If \code{NULL}, then estimated.}

\item{persistenceB}{The persistence vector \eqn{g}, containing smoothing
parameters used in the model B. If \code{NULL}, then estimated.}

\item{initialA}{Either \code{"o"} - optimal or the vector of initials for the
level and / or trend for the model A.}

\item{initialB}{Either \code{"o"} - optimal or the vector of initials for the
level and / or trend for the model B.}

\item{initialSeasonA}{The vector of the initial seasonal components for the
model A. If \code{NULL}, then it is estimated.}

\item{initialSeasonB}{The vector of the initial seasonal components for the
model B. If \code{NULL}, then it is estimated.}

\item{ic}{Information criteria to use in case of model selection.}

\item{h}{Forecast horizon.}

\item{holdout}{If \code{TRUE}, holdout sample of size \code{h} is taken from
the end of the data.}

\item{intervals}{Type of intervals to construct. This can be:

\itemize{
\item \code{none}, aka \code{n} - do not produce prediction
intervals.
\item \code{parametric}, \code{p} - use state-space structure of ETS. In
case of mixed models this is done using simulations, which may take longer
time than for the pure additive and pure multiplicative models.
\item \code{semiparametric}, \code{sp} - intervals based on covariance
matrix of 1 to h steps ahead errors and assumption of normal / log-normal
distribution (depending on error type).
\item \code{nonparametric}, \code{np} - intervals based on values from a
quantile regression on error matrix (see Taylor and Bunn, 1999). The model
used in this process is e[j] = a j^b, where j=1,..,h.
}
The parameter also accepts \code{TRUE} and \code{FALSE}. The former means that
parametric intervals are constructed, while the latter is equivalent to
\code{none}.}

\item{level}{Confidence level. Defines width of prediction interval.}

\item{bounds}{What type of bounds to use in the model estimation. The first
letter can be used instead of the whole word.}

\item{silent}{If \code{silent="none"}, then nothing is silent, everything is
printed out and drawn. \code{silent="all"} means that nothing is produced or
drawn (except for warnings). In case of \code{silent="graph"}, no graph is
produced. If \code{silent="legend"}, then legend of the graph is skipped.
And finally \code{silent="output"} means that nothing is printed out in the
console, but the graph is produced. \code{silent} also accepts \code{TRUE}
and \code{FALSE}. In this case \code{silent=TRUE} is equivalent to
\code{silent="all"}, while \code{silent=FALSE} is equivalent to
\code{silent="none"}. The parameter also accepts first letter of words ("n",
"a", "g", "l", "o").}

\item{xregA}{The vector or the matrix of exogenous variables, explaining some parts
of occurrence variable of the model A.}

\item{xregB}{The vector or the matrix of exogenous variables, explaining some parts
of occurrence variable of the model B.}

\item{xregDo}{Variable defines what to do with the provided xreg:
\code{"use"} means that all of the data should be used, while
\code{"select"} means that a selection using \code{ic} should be done.
\code{"combine"} will be available at some point in future...}

\item{updateX}{If \code{TRUE}, transition matrix for exogenous variables is
estimated, introducing non-linear interactions between parameters.
Prerequisite - non-NULL \code{xreg}.}

\item{...}{The parameters passed to the optimiser, such as \code{maxeval},
\code{xtol_rel}, \code{algorithm} and \code{print_level}. The description of
these is printed out by \code{nloptr.print.options()} function from the \code{nloptr}
package. The default values in the oes function are \code{maxeval=500},
\code{xtol_rel=1E-8}, \code{algorithm="NLOPT_LN_SBPLX"} and \code{print_level=0}.}
}
\value{
The object of class "occurrence" is returned. It contains following list of
values:

\itemize{
\item \code{model} - the type of the estimated ETS model;
\item \code{fitted} - the fitted values for the probability;
\item \code{fittedBeta} - the fitted values of the underlying ETS model, where applicable
(only for occurrence=c("o","i","p"));
\item \code{forecast} - the forecast of the probability for \code{h} observations ahead;
\item \code{forecastBeta} - the forecast of the underlying ETS model, where applicable
(only for occurrence=c("o","i","p"));
\item \code{states} - the values of the state vector;
\item \code{logLik} - the log-likelihood value of the model;
\item \code{nParam} - the number of parameters in the model (the matrix is returned);
\item \code{residuals} - the residuals of the model;
\item \code{actuals} - actual values of occurrence (zeros and ones).
\item \code{persistence} - the vector of smoothing parameters;
\item \code{phi} - the value of the damped trend parameter;
\item \code{initial} - initial values of the state vector;
\item \code{initialSeason} - the matrix of initials seasonal states;
\item \code{occurrence} - the type of the occurrence model.
}

The object of class "occurrence" is returned. It contains following list of
values:

\itemize{
\item \code{model} - the type of the estimated ETS model;
\item \code{fitted} - fitted values of the constructed model;
\item \code{forecast} - forecast for \code{h} observations ahead;
\item \code{states} - values of states (currently level only);
\item \code{logLik} - likelihood value for the model
\item \code{nParam} - number of parameters used in the model;
\item \code{residuals} - residuals of the model;
\item \code{actuals} - actual values of probabilities (zeros and ones).
\item \code{persistence} - the vector of smoothing parameters;
\item \code{initial} - initial values of the state vector;
\item \code{initialSeason} - the matrix of initials seasonal states;
\item \code{occurrence} - the type of the occurrence model.
}
}
\description{
Function returns the occurrence part of iETS model with the specified
probability update and model types.

Function returns the general occurrence model of the of iETS model.
}
\details{
The function estimates probability of demand occurrence, using the selected
ETS state space models. Although the function accepts all types of ETS models,
only the pure multiplicative models make sense.

The function estimates probability of demand occurrence, based on the iETS_G
state-space model. It involves the estimation and modelling of the two
simultaneous state space equations. Thus two parts for the model type,
persistence, intials etc. Although the function accepts all types
of ETS models, only the pure multiplicative models make sense.

The model is based on:

o_t \sim Bernoulli(p_t)
p_t = \frac{a_t}{a_t+b_t},

where a_t and b_t are the parameters of the Beta distribution and are modelled
using separate ETS models.
}
\examples{

y <- rpois(100,0.1)
oes(y, occurrence="o")

oes(y, occurrence="f")


y <- rpois(100,0.1)
oes(y, occurrence="o")

oes(y, occurrence="f")

}
\references{
\itemize{
\item Svetunkov Ivan and Boylan John E. (2017). Multiplicative
State-Space Models for Intermittent Time Series. Working Paper of
Department of Management Science, Lancaster University, 2017:4 , 1-43.
\item Teunter R., Syntetos A., Babai Z. (2011). Intermittent demand:
Linking forecasting to inventory obsolescence. European Journal of
Operational Research, 214, 606-615.
\item Croston, J. (1972) Forecasting and stock control for intermittent
demands. Operational Research Quarterly, 23(3), 289-303.
\item Syntetos, A., Boylan J. (2005) The accuracy of intermittent demand
estimates. International Journal of Forecasting, 21(2), 303-314.
}
}
\seealso{
\code{\link[forecast]{ets}, \link[forecast]{forecast},
\link[smooth]{es}}

\code{\link[forecast]{ets}, \link[forecast]{forecast},
\link[smooth]{es}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}

Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{demand}
\keyword{exponential}
\keyword{forecasting}
\keyword{intermittent}
\keyword{iss}
\keyword{model}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{smooth}
\keyword{smoothing}
\keyword{space}
\keyword{state}
\keyword{ts}
\keyword{univar}
