% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pointLik}
\alias{pointLik}
\alias{pointLik.default}
\title{Point likelihood values}
\usage{
pointLik(object, ...)
}
\arguments{
\item{object}{Time series model.}

\item{...}{Some stuff.}
}
\value{
This function returns a vector.
}
\description{
This function returns a vector of logarithms of likelihoods for each observation
}
\details{
Instead of taking the expected log-likelihood for the whole series, this function
calculates the individual value for each separate observation. Note that these
values are biased, so you would possibly need to take number of degrees of freedom
into account in order to have an unbiased estimator.
}
\examples{

ourModel <- ces(rnorm(100,0,1),h=10)

pointLik(ourModel)

# Bias correction
pointLik(ourModel) - nParam(ourModel)

# Bias correction in AIC style
2*(nParam(ourModel) - pointLik(ourModel))

# BIC calculation based on pointLik
log(nobs(ourModel))*nParam(ourModel) - 2*sum(pointLik(ourModel))

}
\seealso{
\link[stats]{AIC}, \link[stats]{BIC}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{htest}
