\name{sowhat}
\alias{sowhat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function returns the ultimate answer to any question
}
\description{
You need description? So what?
}
\usage{
sowhat(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{ Any number of variables or string with a question.
}
}
\details{
You need details? So what?
}
\value{
It doesn't return any value, only messages. So what?
}
\references{
    \itemize{
        \item{\href{http://en.wiktionary.org/wiki/so_what}{So what?}}
        \item{\href{https://www.youtube.com/watch?v=FJfFZqTlWrQ}{So what?}}
        \item{\href{http://en.wikipedia.org/wiki/Douglas_Adams}{42}}
    }
}
\author{
Ivan Svetunkov
}

\seealso{
Nowwhat (to be implemented), 
}
\examples{
x <- rnorm(10000,0,1);
sowhat(x);

sowhat("What's the meaning of life?")

sowhat("I don't have a girlfriend.")
}
\keyword{ sowhat }
\keyword{ 42 }
